backend/requirements.txt
mysqlclient
django
djangorestframework
django-cors-headers
jinja2
